Enhancement/single activity (#18)
* Remove all activities and dagger-android

+++ app/src/main/java/com/pechuro/bsuirschedule/common/OneFirePerIntervalClickListener.kt
@@ -3,14 +3,14 @@ package com.pechuro.bsuirschedule.common
 import android.os.SystemClock
 import android.view.View
 
-enum class ClickInterval(val milliseconds: Int) {
+enum class ClickInterval(val milliseconds: Long) {
     SHORT(200),
     NORMAL(500),
     LONG(1000)
 }
 
 class OneFirePerIntervalClickListener(
-        private var interval: Int,
+        private var interval: Long,
         private val onSafeCLick: (View) -> Unit
 ) : View.OnClickListener {
